pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mortgage-monitor"
version = "0.2.0"
description = "Comprehensive housing market analysis dashboard with mortgage rate monitoring"
authors = [{name = "Jong-Myun Moon"}]
requires-python = ">=3.13"
readme = "README.md"
license = {file = "LICENSE"}

dependencies = [
    "python-dotenv>=1.0.0",
    "fredapi",
    "pandas",
    "plotly",
    "requests",
    "kaleido>=1.1.0",
    "SQLAlchemy>=2.0.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.1",
    "ruff>=0.13.0",
]

[project.scripts]
mortgage-monitor = "mortgage_monitor.app:main"
macro-dashboard = "macro_dashboard.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mortgage_monitor", "src/macro_dashboard"]
